Fix showing menu when not hovered but menu is open (Issue #16)
Kvantum contextmenu shadows cover the "menubar" causing the titlebar
to no longer be hovered. It probably also sends a "HoverExit" event to
the titlebar as well since Breeze works fine when hovering the
contextmenu.

https://github.com/Zren/material-decoration/issues/16#issuecomment-781460103

This bug is likely a symptom of Issue #29 where Qt treats the QMenu as
a contextmenu not a dropdown menu in a titlebar, however that is
speculation.