Fix repainting the titlebar now that we have borders (Issue #41)
We were getting a slim sliver in the right side of the titlebar that
was not repainting when the switching between inactive/active bg color
due to older repaintTitlebar() code not taking it into account the
titlebar is wider than the client window width when there's borders.

We could probably replace all titleBarRect() calls with calls to
Decoration::titleBar() now that I know about it.