Fix #66 build error on KWindowSystem signal deprecation
cmake VERSION_GREATER_EQUAL requires cmake 3.7, though we already
require the much higher version 3.16.
https://cmake.org/cmake/help/latest/manual/cmake-generator-expressions.7.html?highlight=version_greater_equal#version-comparisons

I can't seem to get it to assign a true/false using set() for some
reason so I has to use and if/else. I was able to test that the old
code still works by changing it to VERSION_LESS and commenting out the
add_definitions (-Wall -Werror) line.

It seems KWindowSystem was imported in a header instead of in various
class body files so I made the import more explicit in those files.